Closes #61
